corrections - 01-06-2026
Only commit

+++ verificationFile/verification_fam_delete.php
@@ -3,10 +3,17 @@ include '../ajaxconfig.php';
 
 $id = $_POST['famid'];
 
-$delct = $connect->query("DELETE FROM `verification_family_info` WHERE id = '$id' ");
+$Qry = $connect->query("SELECT guarentor_name FROM customer_profile WHERE guarentor_name = '$id'");
 
-if ($delct) {
-	$message = " Family Info Deleted Successfully";
+if($Qry->rowCount() > 0){
+	$message = "Family member used as Guarantor.";
+
+}else{
+	$delct = $connect->query("DELETE FROM `verification_family_info` WHERE id = '$id' ");
+	
+	if ($delct) {
+		$message = " Family Info Deleted Successfully";
+	}
 }
 
 echo json_encode($message);
